% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucurve.R
\name{ucurve}
\alias{ucurve}
\title{U-curve algorith on the Partition Lattice Learning Space}
\usage{
ucurve(xtrain, ytrain, xval, yval)
}
\arguments{
\item{xtrain}{Vector with the training sample of x values.}

\item{ytrain}{Vector with the training sample of y values.}

\item{xval}{Vector with the validation sample of x values.}

\item{yval}{Vector with the validation sample of y values.}
}
\value{
\item{hypotheses}{The estimated hypothesis of the global minimums with least VC dimension.}

\item{partitions}{Partitions of the global minimums with least VC dimension.}

\item{error}{Error of the global minimums.}

\item{visited}{Number and percentage of nodes visited.}
}
\description{
TBD
}
\details{
TBD
}
\examples{
set.seed(1)
x <- sample(x = 1:10,size = 50,replace = T)
y <- as.factor(ifelse(x-5+rnorm(50,0,10) > 0,1,0))
x <- factor(x)
train <- sample(1:50,35,F)
xtrain <- x[train]
ytrain <- y[train]
xval <- x[!(c(1:50) \%in\% train)]
yval <- y[!(c(1:50) \%in\% train)]
ucurve(xtrain,ytrain,xval,yval)
}
